moved BinaryFile to io module, refactored out register_binary_to_ref() and apply_shifts_to_binary() to main.
