+++ suite2p/io/__init__.py
@@ -3,3 +3,4 @@ from .nwb import save_nwb, read_nwb
 from .save import save_combined
 from .sbx import sbx_to_binary
 from .tiff import mesoscan_to_binary, ome_to_binary, tiff_to_binary, generate_tiff_filename, save_tiff
+from .binary import BinaryFile
