+++ suite2p/registration/register.py
@@ -133,143 +133,6 @@ def compute_crop(xoff, yoff, corrXY, th_badframes, badframes, maxregshift, Ly, L
     return badframes, yrange, xrange
 
 
-class BinaryFile:
-
-    def __init__(self, nbatch: int, Ly: int, Lx: int, nframes: int, reg_file: str, raw_file: str):
-        self.nbatch = nbatch
-        self.Ly = Ly
-        self.Lx = Lx
-        self.nframes = nframes
-        self.reg_file = open(reg_file, mode='wb' if raw_file else 'r+b')
-        self.raw_file = open(raw_file, 'rb') if raw_file else None
-
-        self._nfr = 0
-        self._can_read = True
-
-    @property
-    def nbytesread(self):
-        return 2 * self.Ly * self.Lx * self.nbatch
-
-    def close(self):
-        self.reg_file.close()
-        if self.raw_file:
-            self.raw_file.close()
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.close()
-
-    def __iter__(self):
-        return self
-
-    def __next__(self):
-        data = self.read()
-        if data is None:
-            raise StopIteration
-        return data
-
-    def read(self, dtype=np.float32) -> Optional[np.ndarray]:
-        if not self._can_read:
-            raise IOError("BinaryFile needs to write before it can read again.")
-        buff = self.raw_file.read(self.nbytesread) if self.raw_file else self.reg_file.read(self.nbytesread)
-        data = np.frombuffer(buff, dtype=np.int16, offset=0).reshape(-1, self.Ly, self.Lx).astype(dtype)
-        if (data.size == 0) | (self._nfr >= self.nframes):
-            return None
-        self._nfr += data.size
-        self._can_read = False
-        return data
-
-    def write(self, data):
-        if self._can_read:
-            raise IOError("BinaryFile needs to read before it can write again.")
-
-        if not self.raw_file:
-            self.reg_file.seek(-2 * data.size, 1)
-        self.reg_file.write(bytearray(np.minimum(data, 2 ** 15 - 2).astype('int16')))
-        self._can_read = True
-
-
-def register_binary_to_ref(nbatch: int, Ly: int, Lx: int, nframes: int, ops, refAndMasks, reg_file_align, raw_file_align):
-
-    with BinaryFile(nbatch=nbatch, Ly=Ly, Lx=Lx, nframes=nframes, reg_file=reg_file_align, raw_file=raw_file_align) as f:
-        for data in f:
-
-            data, ymax, xmax, cmax, yxnr = compute_motion_and_shift(
-                data=data,
-                bidiphase=ops['bidiphase'],
-                bidi_corrected=ops['bidi_corrected'],
-                refAndMasks=refAndMasks,
-                maxregshift=ops['maxregshift'],
-                nblocks=ops['nblocks'],
-                xblock=ops['xblock'],
-                yblock=ops['yblock'],
-                nr_sm=ops['NRsm'],
-                snr_thresh=ops['snr_thresh'],
-                smooth_sigma_time=ops['smooth_sigma_time'],
-                maxregshiftNR=ops['maxregshiftNR'],
-                is_nonrigid=ops['nonrigid'],
-                reg_1p=ops['1Preg'],
-                spatial_hp=ops['spatial_hp'],
-                pre_smooth=ops['pre_smooth'],
-            )
-
-            # output
-            rigid_offset = [ymax, xmax, cmax]
-            nonrigid_offset = list(yxnr)
-            yield rigid_offset, nonrigid_offset, data
-
-            f.write(data)
-
-
-
-
-def apply_shifts_to_binary(batch_size: int, Ly: int, Lx: int, nframes: int,
-    is_nonrigid: bool, bidiphase_value: int, bidi_corrected, nblocks, xblock, yblock,
-    offsets, reg_file_alt, raw_file_alt):
-    """ apply registration shifts computed on one binary file to another
-    
-    Parameters
-    ----------
-
-    offsets : list of arrays
-        shifts computed from reg_file_align/raw_file_align, 
-        rigid shifts in Y are offsets[0] and in X are offsets[1], 
-        nonrigid shifts in Y are offsets[3] and in X are offsets[4]
-
-    reg_file_alt : string
-        file to (read if raw_file_align empty, and) write registered binary to
-
-    raw_file_align : string
-        file to read raw binary from (if not empty)
-    """
-
-    nfr = 0
-    with BinaryFile(nbatch=batch_size, Ly=Ly, Lx=Lx, nframes=nframes, reg_file=reg_file_alt, raw_file=raw_file_alt) as f:
-        for data in f:
-
-            # get shifts
-            nframes = data.shape[0]
-            iframes = nfr + np.arange(0, nframes, 1, int)
-            ymax, xmax = offsets[0][iframes].astype(np.int32), offsets[1][iframes].astype(np.int32)
-
-            # apply shifts
-            if bidiphase_value != 0 and not bidi_corrected:
-                bidiphase.shift(data, bidiphase_value)
-
-            rigid.shift_data(data, ymax, xmax)
-
-            if is_nonrigid:
-                ymax1, xmax1 = offsets[3][iframes], offsets[4][iframes]
-                data = nonrigid.transform_data(data, nblocks=nblocks, xblock=xblock, yblock=yblock, ymax1=ymax1, xmax1=xmax1)
-
-            yield data
-            nfr += nframes
-
-            f.write(data)
-
-
 def pick_initial_reference(frames):
     """ computes the initial reference image
